+++ src/App.tsx
@@ -8,9 +8,17 @@ import Oauth from "./pages/Oauth/Oauth";
 import GithubOauth from "./pages/Oauth/GithubOAuth";
 import { ResetStyle, GlobalStyle } from "./components/globalStyle.js";
 import { Navigate, Route, Routes } from "react-router-dom";
+import { SelectContext } from "./utils/SelectContext";
+import { RepoLabelArr } from "../src/type";
+
+import { useState } from "react";
 const REPOSITORY = "github-project";
 
 function App() {
+  const [RepoLabelArr, setRepoLabelArr] = useState<RepoLabelArr>(
+    [] as unknown as RepoLabelArr
+  );
+  const [selectedEdit, setSelectedEdit] = useState<number>();
   return (
     <>
       <ResetStyle />
@@ -22,7 +30,14 @@ function App() {
           element={<GithubOauth />}
         ></Route>
         <Route path={`/${REPOSITORY}/ooath`} element={<Oauth />}></Route>
-        <Route path={`/${REPOSITORY}`} element={<Label />}></Route>
+        <SelectContext.Provider
+          value={{
+            RepoLabelArr: [RepoLabelArr, setRepoLabelArr],
+            selectedEdit: [selectedEdit, setSelectedEdit],
+          }}
+        >
+          <Route path={`/${REPOSITORY}`} element={<Label />}></Route>
+        </SelectContext.Provider>
         <Route
           path={`/${REPOSITORY}/issuelist`}
           element={<IssueList />}
