+++ src/type.tsx
@@ -1,0 +1,10 @@
+export type RepoLabelData = {
+  color?: string;
+  default?: boolean;
+  description?: string;
+  id?: number;
+  name?: string;
+  node_id?: number;
+  url?: string;
+};
+export type RepoLabelArr = [RepoLabelData];
