+++ src/utils/SelectContext.tsx
@@ -1,0 +1,8 @@
+import { createContext, Dispatch, SetStateAction } from "react";
+import { RepoLabelArr } from "../type";
+
+type ContextState = {
+  RepoLabelArr: [RepoLabelArr, Dispatch<SetStateAction<RepoLabelArr>>];
+  selectedEdit: [Number, Dispatch<SetStateAction<Number>>];
+};
+export const SelectContext = createContext({} as ContextState);
